mini-porjects/03-medallion-architecture-with-spark-and-hadoop/orchestrator/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
python-dotenv==1.0.1
requests==2.32.3
setuptools==69.5.1
urllib3==2.2.1
wheel==0.43.0
